+++ packages/@sanity/plugin-kit/src/actions/inject.ts
@@ -1,7 +1,6 @@
 import path from 'path'
 import {fileURLToPath} from 'url'
 
-// @ts-expect-error missing types
 import licenses from '@rexxars/choosealicense-list'
 import gitRemoteOriginUrl from 'git-remote-origin-url'
 
@@ -194,9 +193,9 @@ async function getLicense(
   }
 
   const text = license.body
-    .replace(/\[fullname\]/g, user?.name)
-    .replace(/\[project\]/g, pluginName)
-    .replace(/\[year\]/g, new Date().getFullYear())
+    .replace(/\[fullname\]/g, user?.name ?? '')
+    .replace(/\[project\]/g, pluginName ?? '')
+    .replace(/\[year\]/g, String(new Date().getFullYear()))
 
   return {id: license.id, text}
 }
@@ -325,8 +324,6 @@ async function writeStaticAssets(options: InjectOptions) {
     flags.eslint && eslintrcTemplate({flags: options.flags}),
     flags.eslint && eslintignoreTemplate({outDir, flags: options.flags}),
     {type: 'copy', from: 'editorconfig', to: '.editorconfig'},
-    {type: 'copy', from: 'sanity.json', to: 'sanity.json'},
-    {type: 'copy', from: 'v2-incompatible.js.template', to: 'v2-incompatible.js'},
     pkgConfigTemplate({outDir, flags: options.flags}),
     flags.gitignore && gitignoreTemplate(),
     flags.typescript && tsconfigTemplate({flags: options.flags}),