+++ plugins/sanity-plugin-mux-input/package.config.ts
@@ -4,5 +4,4 @@ import {defineConfig} from '@sanity/pkg-utils'
 export default defineConfig({
   ...config,
   babel: {styledComponents: true},
-  external: (prev) => prev.filter((dep) => dep !== 'use-error-boundary'),
 })