+++ plugins/sanity-plugin-dashboard-widget-vercel/CHANGELOG.md
@@ -1,5 +1,27 @@
 # sanity-plugin-dashboard-widget-vercel
 
+## 4.0.8
+
+### Patch Changes
+
+- [#1424](https://github.com/sanity-io/plugins/pull/1424) [`56ef1aa`](https://github.com/sanity-io/plugins/commit/56ef1aa2134022b509edcd225dcf6775a82ea359) Thanks [@squiggler-app](https://github.com/apps/squiggler-app)! - fix(deps): update dependency @tanstack/react-query to ^5.101.1
+
+- [#1437](https://github.com/sanity-io/plugins/pull/1437) [`9786ded`](https://github.com/sanity-io/plugins/commit/9786ded76ef680abb4a4d9827eae7059cceb7400) Thanks [@squiggler-app](https://github.com/apps/squiggler-app)! - fix(deps): update dependency xstate to ^5.32.2
+
+## 4.0.7
+
+### Patch Changes
+
+- [#1396](https://github.com/sanity-io/plugins/pull/1396) [`6e30c3c`](https://github.com/sanity-io/plugins/commit/6e30c3c6d454cb05e08b23dcc7cba7c8b7c8e72b) Thanks [@squiggler-app](https://github.com/apps/squiggler-app)! - fix(deps): update dependency react-hook-form to ^7.80.0
+
+## 4.0.6
+
+### Patch Changes
+
+- [#1376](https://github.com/sanity-io/plugins/pull/1376) [`ca14e36`](https://github.com/sanity-io/plugins/commit/ca14e36eb6d5cb1bffdc2d58393b8a39456c5dd2) Thanks [@squiggler-app](https://github.com/apps/squiggler-app)! - fix(deps): Update tanstack-query monorepo
+
+- [#1373](https://github.com/sanity-io/plugins/pull/1373) [`0f2e85c`](https://github.com/sanity-io/plugins/commit/0f2e85c3ccaf086efb2715ad5b9211bdc6d9c1cf) Thanks [@stipsan](https://github.com/stipsan)! - Remove direct `javascript-time-ago` dependency now that `react-time-ago` bundles it
+
 ## 4.0.5
 
 ### Patch Changes